fixed type conversion issue (changed x,y from float to int in item and estusflask)

+++ Model/EstusFlask.cpp
@@ -1,7 +1,7 @@
 #include "EstusFlask.h"
 
 EstusFlask::EstusFlask(std::string itemName, int itemPrice, std::string itemDescription,
-                       int maxItemCount, int itemCount, float x, float y, int width, int height): 
+                       int maxItemCount, int itemCount, int x, int y, int width, int height): 
                        Item(itemName, itemPrice, itemDescription, maxItemCount, itemCount, x, y, width, height){
 
 }
